requirements.txt
fastapi
uvicorn
langchain
pydub
openai
python-dotenv
langchain_community
